(minor) Move infer_xlabs_ip_adapter_params_from_state_dict(...) to state_dict_utils.py.
