+++ tests/backend/flux/ip_adapter/test_xlabs_ip_adapter_flux.py
@@ -1,10 +1,12 @@
 import accelerate
 import torch
 
-from invokeai.backend.flux.ip_adapter.state_dict_utils import is_state_dict_xlabs_ip_adapter
+from invokeai.backend.flux.ip_adapter.state_dict_utils import (
+    infer_xlabs_ip_adapter_params_from_state_dict,
+    is_state_dict_xlabs_ip_adapter,
+)
 from invokeai.backend.flux.ip_adapter.xlabs_ip_adapter_flux import (
     XlabsIpAdapterFlux,
-    infer_xlabs_ip_adapter_params_from_state_dict,
 )
 from tests.backend.flux.ip_adapter.xlabs_flux_ip_adapter_state_dict import xlabs_sd_shapes
